requirements.txt
pyyaml>=6.0
maxminddb>=2.0
